Created lot_data_with_capacity.csv
For each parking lot on North Campus, lot_data_with_capacity will have the type of lot (faculty & staff, students, or commuter (inclusive or)), the count of the "regular" spots, handicap spots, Zipcar spots, and of course, the total capacity. It is important to have a distribution of that max capacity.