Specified the type of lot for each North Campus parking lot
A 1 bit under Faculty and Staff means Faculty and Staff can park in the lot. Similarly, a 1 bit under Students or Commuter means students or commuters can park in that lot, respectively. If a bit under Commuter is a 1, then that means the only kinds of students that can park in that lot are commuter students, hence the bit under Students would simultaneously be 0.